Answer which position is left nothing from the rules, not from what survived
Review found that a reading admitting nothing blames a different position
depending on the order its rules were met in. It is worse than an order: one of
the answers is wrong.

    X = a == 0 && b == 0
    Y = a == 0
    Z = (a == 1 && b == 0) || (a == 0 && b == 1)

    X ∧ Y ∧ Z    b is left 0
    Y ∧ Z ∧ X    b is left nothing

Nothing satisfies the three either way. Read one position at a time they leave `a`
at 0 and `b` at 0, so neither is a position the rules leave nothing at — what they
cannot do is hold together. The second answer sends an author to a rule about `b`
that is not the reason.

**It is not a bug in when dead pairs are dropped.** Dropping them is right: a pair
nothing stands in is not an alternative. The mistake was asking the alternatives a
question they cannot answer. A conjunction meets them pairwise and drops what dies,
so what a dropped pair was going to say about a position leaves with it, and any
answer read off the survivors follows the order.

So the reading carries what each position's own rules leave it, beside the
alternatives and by the same connectives. Not what a position may hold — `at` is
narrower wherever the alternatives are held apart, which is the whole point of
holding them — and not a second account of that. It answers the one question the
alternatives cannot, and answers it the same however the rules were bracketed,
because a meet of these is their meet at each position.

`BottomResidue` and `PairOutcome` go with it: what was being reconstructed at the
moment of death is now simply there, and telling a live product from a dead one is
a box refusing to be built with an empty side — which review also asked for, and
which is what lets a reading say it admits nothing by being `Nothing` and nothing
else.

**A count is answered in a wider type than the limit is written in.** Review found
that a limit at the top of `int` makes the ceiling the smallest number there is, so
every count is under it and the guardrail admits exactly what it exists to refuse —
at the one limit somebody reaching for "no limit" would write. Measured with the
ceiling counted in the limit's own type:

    one past the limit, whatever the limit is: 2147483647
      ==> expected: <2147483648> but was: <-2147483648>

And a limit below one is refused where it is written: a reading holds at least one
alternative, so a bound under that bounds nothing, and a resource bound that admits
everything is the absence of one.

`takingValuesRead` says the values of a state once. It can no longer combine two
readings, and the one way left to get it wrong is saying it twice, which would keep
the second and drop the first without a word.

+++ souther-compiler/src/main/java/souther/compiler/check/ExpansionCost.java
@@ -30,13 +30,18 @@ import java.util.List;
  * wanted — what is asked of it is whether the limit is exceeded — and a conjunction of choices
  * leaves the range of a count long before it leaves the range of what an author can write.
  */
-final class ExpansionCost implements ClauseReading<Integer> {
+final class ExpansionCost implements ClauseReading<Long> {
 
-    /** One past the limit, which is where the arithmetic stops. */
-    private final int ceiling;
+    /** One past the limit, which is where the arithmetic stops.
+     *
+     *  <p>Counted in a wider type than the limit is written in, and answered in one. Counted in the
+     *  limit's own type, one past the largest limit is the smallest number there is and every count
+     *  is under it — so the guardrail would admit exactly the expansions it exists to refuse, and
+     *  would do it at the one limit somebody reaching for "no limit" would write. */
+    private final long ceiling;
 
     ExpansionCost(int limit) {
-        this.ceiling = limit + 1;
+        this.ceiling = (long) limit + 1;
     }
 
     /**
@@ -45,9 +50,9 @@ final class ExpansionCost implements ClauseReading<Integer> {
      *
      * @param limit the count is asked about, saturating at one past it
      */
-    static int of(List<Core> clauses, int limit) {
+    static long of(List<Core> clauses, int limit) {
         ExpansionCost counting = new ExpansionCost(limit);
-        int cost = counting.nothingSaid();
+        long cost = counting.nothingSaid();
         for (Core clause : clauses) {
             cost = counting.both(cost, counting.read(clause, true));
         }
@@ -56,8 +61,8 @@ final class ExpansionCost implements ClauseReading<Integer> {
 
     /** Nothing read is one alternative — the empty product — and is the identity of the fold. */
     @Override
-    public Integer nothingSaid() {
-        return 1;
+    public Long nothingSaid() {
+        return 1L;
     }
 
     /**
@@ -68,17 +73,17 @@ final class ExpansionCost implements ClauseReading<Integer> {
      * leaves come out costing nothing and be admitted under any budget.
      */
     @Override
-    public Integer leaf(Core e, boolean positive) {
-        return 1;
+    public Long leaf(Core e, boolean positive) {
+        return 1L;
     }
 
     @Override
-    public Integer both(Integer one, Integer other) {
-        return (int) Math.min(ceiling, (long) one * other);
+    public Long both(Long one, Long other) {
+        return Math.min(ceiling, one * other);
     }
 
     @Override
-    public Integer either(Integer one, Integer other) {
+    public Long either(Long one, Long other) {
         return Math.min(ceiling, one + other);
     }
 }
